web/requirements.txt
boto==2.38.0
dj-database-url==0.3.0
Django==1.8.3
django-filter==0.10.0
django-storages-redux==1.3
django-suit==0.2.14
djangorestframework==3.2.1
gunicorn==19.3.0
loremipsum==1.0.5
Markdown==2.6.2
mock==1.3.0
pbr==1.5.0
Pillow==2.9.0
psycopg2==2.6.1
requests==2.7.0
six==1.9.0
testfixtures==4.2.0
